created more sensible components
